requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
cattrs==24.1.2
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
coremltools==8.2
cycler==0.12.1
datasets==3.4.0
debugpy==1.8.13
decorator==5.2.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.34.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
netron==8.2.1
networkx==3.4.2
numpy==1.24.3
onnx==1.17.0
onnxruntime==1.21.0
onnxruntime-silicon==1.16.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==6.30.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pyarrow==19.0.1
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.30.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
